refactor(widgets): create classes.ts file and move BreakpointChecker there
