+++ packages/widgets/src/lib/index.ts
@@ -15,12 +15,3 @@ export { PagesNavigator } from './components/pages-navigator/index.js';
 export { ScrollToTopButton } from './components/scroll-to-top/index.js';
 
 export { TOC } from './components/toc/index.js';
-
-// `ScreenSizes` contains the pixel values for different screen sizes.
-export const ScreenSizes = {
-	xs: 320,
-	sm: 640,
-	md: 768,
-	lg: 1024,
-	xl: 1280
-};
