+++ packages/widgets/src/lib/utils.ts
@@ -1,5 +1,5 @@
 import { isNullish } from '@sveltinio/ts-utils/is';
-import type { BreakpointMatch, ResourceContent, Screens } from './types';
+import type { ResourceContent } from './types';
 import { contains } from '@sveltinio/ts-utils/collections';
 
 /**
@@ -24,113 +24,6 @@ export function retrieveCssClassNames(value: any, reservedNames: string[]): stri
 	return value;
 }
 
-/* The BreakpointChecker class provides methods to check the screen size and breakpoints of a device. */
-export class BreakpointChecker {
-	private _screens: Screens;
-	private _mobileSize: number;
-	private _desktopSize: number;
-
-	private _defaultScreenSizes: Screens = {
-		xs: { min: '320', max: '639' },
-		sm: { min: '640px', max: '767px' },
-		md: { min: '768px', max: '1023px' },
-		lg: { min: '1024px', max: '1279px' },
-		xl: { min: '1280px', max: '1535px' },
-		'2xl': { min: '1536px' }
-	};
-
-	constructor(theScreens?: Screens);
-	constructor(theScreens?: Screens, mobileSize?: number, desktopSize?: number) {
-		this._screens = theScreens || this._defaultScreenSizes;
-
-		this._mobileSize = mobileSize || 768;
-		this._desktopSize = desktopSize || 1024;
-	}
-
-	public get screens(): Screens {
-		return this._screens;
-	}
-	public set screens(value: Screens) {
-		this._screens = value;
-	}
-
-	public get mobileSize(): number {
-		return this._mobileSize;
-	}
-	public set mobileSize(value: number) {
-		this._mobileSize = value;
-	}
-
-	public get desktopSize(): number {
-		return this._desktopSize;
-	}
-	public set desktopSize(value: number) {
-		this._desktopSize = value;
-	}
-
-	public isMobile = (width: number): boolean => width < this.mobileSize;
-	public isDesktop = (width: number): boolean => width >= this.desktopSize;
-	public isTablet = (width: number): boolean =>
-		width >= this.mobileSize && width < this.desktopSize;
-
-	public isXSmallScreen = (width: number): boolean => this.getBreakpointName(width) === 'xs';
-	public isSmallScreen = (width: number): boolean => this.getBreakpointName(width) === 'sm';
-	public isMediumScreen = (width: number): boolean => this.getBreakpointName(width) === 'md';
-	public isLargeScreen = (width: number): boolean => this.getBreakpointName(width) === 'lg';
-	public isXLargeScreen = (width: number): boolean => this.getBreakpointName(width) === 'xl';
-	public isXXLargeScreen = (width: number): boolean => this.getBreakpointName(width) === '2xl';
-
-	private getBreakpointName(width: number): BreakpointMatch {
-		const breakpointKeys = Object.keys(this._screens) as BreakpointMatch[];
-		let breakpoint: BreakpointMatch = breakpointKeys[0];
-
-		for (let i = 1; i < breakpointKeys.length; i++) {
-			const prevBreakpoint = this._screens[breakpoint];
-			const currBreakpoint = this._screens[breakpointKeys[i]];
-
-			if (
-				width >= parseInt(prevBreakpoint.min) &&
-				width <= parseInt(currBreakpoint.max || '10000')
-			) {
-				breakpoint = breakpointKeys[i];
-			}
-		}
-
-		return breakpoint;
-	}
-}
-
-/**
- * Returns the matching breakpoint for a given width based on predefined screen sizes.
- *
- * @param width - The width of the screen in pixels, used to determine which breakpoint the
- * screen falls into.
- * @returns a string representing the current breakpoint based on the input width value.
- */
-export function getBreakpoint(width: number): BreakpointMatch {
-	const screens: Screens = {
-		xs: { min: '320', max: '639' },
-		sm: { min: '640px', max: '767px' },
-		md: { min: '768px', max: '1023px' },
-		lg: { min: '1024px', max: '1279px' },
-		xl: { min: '1280px', max: '1535px' },
-		'2xl': { min: '1536px' }
-	};
-	const breakpointKeys = Object.keys(screens) as BreakpointMatch[];
-	let breakpoint: BreakpointMatch = breakpointKeys[0];
-
-	for (let i = 1; i < breakpointKeys.length; i++) {
-		const prevBreakpoint = screens[breakpoint];
-		const currBreakpoint = screens[breakpointKeys[i]];
-
-		if (width >= parseInt(prevBreakpoint.min) && width <= parseInt(currBreakpoint.max || '10000')) {
-			breakpoint = breakpointKeys[i];
-		}
-	}
-
-	return breakpoint;
-}
-
 /**
  * The function checks if a given string is a single character.
  * @param txt - A string to be checked if it is a single character or not.
@@ -153,34 +46,47 @@ export function isChar(txt: string): boolean {
 export function getIdFromAriaLabel(node: HTMLElement): string {
 	let id = node.getAttribute('aria-label');
 	if (id) {
-		id = id.trim().toLowerCase().replace(' ', '-').replace('/', '-');
+		id = id.trim().toLowerCase().replace(/\s/g, '-').replace('/', '-');
 		return id;
 	}
 	return '';
 }
 
-/**
- * Adds a CSS class to an HTML element.
- *
- * @param item - This is the HTML element to which the CSS class will be
- * added.
- * @param className - A string parameter that represents the name of the CSS
- * class that needs to be added to the specified HTMLElement.
- */
-export function addCssClass(item: HTMLElement, className: string) {
-	item.classList.add(className);
+export function getMenuId(node: HTMLElement): string {
+	let id = '';
+	const role = node.getAttribute('role');
+
+	if (role) {
+		id = role + '-' + getIdFromAriaLabel(node);
+	}
+
+	return id;
 }
 
 /**
- * Removes a specified CSS class from a given HTML element.
+ * Adds specified CSS classes from a given HTML element.
  *
- * @param item  - This is the HTML element to which the CSS class needs to
- * be removed.
- * @param className - A string that represents the name of the CSS
- * class that needs to be removed from the specified HTMLElement.
+ * @param item - the HTML element to remove CSS classes from
+ * @param names - An array of strings representing the CSS class names that should be added from
+ * the `classList` property of the `item` parameter.
  */
-export function removeCssClass(item: HTMLElement, className: string) {
-	item.classList.remove(className);
+export function addCssClasses(item: HTMLElement | null, names: string[]) {
+	if (!isNullish(item)) {
+		names.map((n) => item.classList.add(n));
+	}
+}
+
+/**
+ * Removes specified CSS classes from a given HTML element.
+ *
+ * @param item - the HTML element to remove CSS classes from
+ * @param names - An array of strings representing the CSS class names that should be removed from
+ * the `classList` property of the `item` parameter.
+ */
+export function removeCssClasses(item: HTMLElement | null, names: string[]) {
+	if (!isNullish(item)) {
+		names.map((n) => item.classList.remove(n));
+	}
 }
 
 /**
@@ -189,12 +95,11 @@ export function removeCssClass(item: HTMLElement, className: string) {
  * @param items - An array of HTMLElements that need to have their focus reset (i.e.
  * have the CSS class 'focus' removed).
  */
-export function resetFocusClass(items: HTMLElement[]) {
-	items.forEach((item) => removeCssClass(item, 'focus'));
+export function resetFocusClass(items: HTMLElement[] | null) {
+	if (!isNullish(items)) items.forEach((item) => removeCssClasses(item, ['focus']));
 }
-
 /**
- *! The following are used as utilities in the demo pages
+ *! The following are used as utilities for the demo pages
  */
 
 export function getURL(base: string, item: ResourceContent): string {
